+++ cdp-agentkit-core/cdp_agentkit_core/actions/__init__.py
@@ -11,6 +11,7 @@ from cdp_agentkit_core.actions.transfer import TransferAction
 from cdp_agentkit_core.actions.wow.buy_token import WowBuyTokenAction
 from cdp_agentkit_core.actions.wow.create_token import WowCreateTokenAction
 from cdp_agentkit_core.actions.wow.sell_token import WowSellTokenAction
+from cdp_agentkit_core.actions.wrap_eth import WrapEthAction
 
 
 # WARNING: All new CdpAction subclasses must be imported above, otherwise they will not be discovered
@@ -40,4 +41,5 @@ __all__ = [
     "WowBuyTokenAction",
     "WowCreateTokenAction",
     "WowSellTokenAction",
+    "WrapEthAction",
 ]